move InsertSetEncryptValueToken to rewrite package
